[update] filter sample merge
