+++ VisitControlSystem/src/Filter.java
@@ -1,4 +1,3 @@
-package Pipe_Filter;
 public interface Filter {
    public Object execute (Object request);
 }
