+++ VisitControlSystem/src/FilterManager.java
@@ -1,4 +1,3 @@
-package Pipe_Filter;
 public class FilterManager {
    FilterChain filterChain;
