+++ VisitControlSystem/src/HMessage.java
@@ -1,4 +1,3 @@
-package Pipe_Filter;
 import java.util.ArrayList;
 import java.util.List;
