+++ VisitControlSystem/src/PackMessage.java
@@ -1,0 +1,37 @@
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
+public class PackMessage {
+	
+	public PackMessage(Object obj, String content) {
+    	this.obj = obj;
+    	this.content = content;
+    }
+	private Object obj;
+	private String content;
+	
+	public void printPackMessage() {
+		if (obj.getClass() == Visitor.class) {			
+//			System.out.println("Visitor : " + ((Visitor)obj).getVid() + "Content :" + this.content);
+			Visitor visitor = (Visitor)obj;
+			StringBuilder sb = new StringBuilder();
+			sb.append("[코로나 확진판별 검사 안내]\n");
+			sb.append(visitor.getVname() + "님 역학조사 결과 코로나 확진자와 접촉한 것으로 확인되었습니다.\n");
+			sb.append("인근 선별진료소를 통해 확진여부를 검사해주시길 바랍니다.\n");
+			System.out.println(sb.toString());
+		}
+		else if (obj.getClass() == Facility.class) {
+//			System.out.println("Facility : " + ((Facility)obj).getFid() + "Content :" + this.content);
+			Facility visitor = (Facility)obj;
+			StringBuilder sb = new StringBuilder();
+			sb.append("[코로나 확진판별 검사 안내]\n");
+			sb.append("\"" + visitor.getFname() + "\"" + "관리자님 역학조사 결과 코로나 확진자가 해당 시설에 방문한 것으로 확인되었습니다.\n");
+			sb.append("2주간 가게 임시휴업 조치 및 방역조치 부탁드리며, 관리자님 또한 인근 선별진료소를 통해 확진여부를 검사해주시길 바랍니다.\n");
+			System.out.println(sb.toString());
+		}
+		else {
+			System.out.println("error");
+		}		
+	}
+}
