--- VisitControlSystem/src/Pipe_Filter/Facility.java
@@ -1,24 +1,0 @@
-package Pipe_Filter;
-
-public class Facility {
-
-
-    public Facility(Long id, String name, String address, String contact) {
-        this.fid = id;
-        this.fname = name;
-        this.faddress = address;
-        this.fcontact = contact;
-
-    }
-
-    private Long fid;
-    private String fname;
-    private String faddress;
-    private String fcontact;
-    
-    ////
-    public Long getid() {
-    	return this.fid;
-    }
-    ////
-}
