--- VisitControlSystem/src/Pipe_Filter/Vaccination.java
@@ -1,6 +1,0 @@
-package Pipe_Filter;
-
-public enum Vaccination {
-    UNVACCINATED, FIRST_VACCINATION_COMPLETED, SECOND_VACCINATION_COMPLETED
-}
-
