+++ VisitControlSystem/src/Target.java
@@ -1,5 +1,3 @@
-package Pipe_Filter;
-
 import java.util.ArrayList;
 import java.util.List;
